requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.9.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
discord-ext-menus==1.1
discord.py @ git+https://github.com/Rapptz/discord.py@0981f122afc69772a3541b205b382256cc8f3d1d
expiringdict==1.2.2
frozenlist==1.3.0
idna==3.3
lavalink==4.0.1
multidict==6.0.2
mysql-connector-python==8.0.29
Pillow==9.3.0
protobuf==3.20.2
pycparser==2.21
PyNaCl==1.5.0
python-socketio==5.7.2
PyYAML==6.0
requests==2.27.1
StringProgressBar==1.1.1
spotifysearch==0.0.5
swaglyrics==1.2.2
Unidecode==1.1.1
urllib3==1.26.9
yarl==1.7.2
